% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fn_parameters_as_vars.R
\name{assign_fn_parameters_as_vars}
\alias{assign_fn_parameters_as_vars}
\title{Assign function parameters as variables}
\usage{
assign_fn_parameters_as_vars(fun = NULL)
}
\arguments{
\item{fun}{a function}
}
\description{
Take a function and assign all the parameters defined within it
as variables in the global environment
}
\details{
This function can be useful when you are testing a function and
you need to set all the function's parameters in a single operation.
}
\examples{
\dontrun{
assign_fn_parameters_as_vars(pm)
assign_fn_parameters_as_vars(mean)
assign_fn_parameters_as_vars(sd)
kim::assign_fn_parameters_as_vars(floodlight_2_by_continuous)
?get
assign_fn_parameters_as_vars(scatterplot)
traceback()
exists("1")
?exists("mean")
getAnywhere("mean")
environment(mean)
rm(list = ls())
function_as_string <- "floodlight_2_by_continuous"
search()
print(paste0("ls list: ", ls()))
}
}
