% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_for_counts.R
\name{barplot_for_counts}
\alias{barplot_for_counts}
\title{Barplot for counts}
\usage{
barplot_for_counts(data = NULL, x, y)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{x}{name of the variable that will be on the x axis of the barplot}

\item{y}{name of the variable that will be on the y axis of the barplot}
}
\description{
Barplot for counts
}
\examples{
\donttest{
barplot_for_counts(x = 1:3, y = 7:9)
}
\donttest{
barplot_for_counts(data = data.frame(
cyl = names(table(mtcars$cyl)), count = as.vector(table(mtcars$cyl))),
x = "cyl", y = "count")
}
}
