% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R
\name{bracket}
\alias{bracket}
\title{Draw a bracket on a ggplot}
\usage{
bracket(
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  vertical = NULL,
  horizontal = NULL,
  open = NULL,
  bracket_shape = NULL,
  thickness = 2,
  bracket_color = "black",
  label = NULL,
  label_hjust = NULL,
  label_vjust = NULL,
  label_font_size = 5,
  label_font_face = "bold",
  label_color = "black",
  label_parse = FALSE
)
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{vertical}{vertical}

\item{horizontal}{horizontal}

\item{open}{open}

\item{bracket_shape}{bracket_shape}

\item{thickness}{thickness}

\item{bracket_color}{bracket_color}

\item{label}{label}

\item{label_hjust}{label_hjust}

\item{label_vjust}{label_vjust}

\item{label_font_size}{label_font_size}

\item{label_font_face}{label_font_face}

\item{label_color}{label_font_face}

\item{label_parse}{label_parse}
}
\value{
a ggplot object; there will be no meaningful output from
this function. Instead, this function should be used with another
ggplot object
}
\description{
Draw a square bracket with a label on a ggplot
}
\examples{
\donttest{
library(ggplot2)
ggplot(mtcars, aes(x = cyl, y = mpg)) + geom_point() +
bracket(6.1, 6.2, 17, 22, bracket_shape = "]", label = "abc")
}
}
