% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a substring}
\usage{
capitalize(x, start = 1, end = 1)
}
\arguments{
\item{x}{a character string or a character vector}

\item{start}{starting position of the susbtring (default = 1)}

\item{end}{ending position of the susbtring (default = 1)}
}
\value{
a character string or a character vector
}
\description{
Capitalizes the first letter (by default) or a substring
of a given character string or each element of the character vector
}
\examples{
capitalize("abc")
capitalize(c("abc", "xyx"), start = 2, end = 3)
}
