% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_first_letter.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize the first letter}
\usage{
capitalize_first_letter(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string
}
\description{
Capitalizes the first letter of a given character string
}
\examples{
capitalize_first_letter("abc")
}
