% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_modes.R
\name{check_modes}
\alias{check_modes}
\title{Check modes of objects}
\usage{
check_modes(..., mode_to_confirm = NULL)
}
\arguments{
\item{...}{R objects.}

\item{mode_to_confirm}{The function will test whether each input is
of this mode. For example, \code{
check_modes(a, mode_to_confirm = "numeric")}, the function will check
whether the object \code{a} is numeric.}
}
\description{
Check modes of objects
}
\examples{
check_modes(1L, mode_to_confirm = "numeric")
check_modes(
TRUE, FALSE, 1L, 1:3, 1.1, c(1.2, 1.3), "abc", 1 + 2i, intToBits(1L),
mode_to_confirm = "numeric")
}
