% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_req_pkg.R
\name{check_req_pkg}
\alias{check_req_pkg}
\title{Check for required packages}
\usage{
check_req_pkg(pkg = NULL)
}
\arguments{
\item{pkg}{a character vector containing names of packages to check}
}
\value{
there will be no output from this function. Rather, the
function will check whether the packages given as inputs are installed.
}
\description{
Check whether required packages are installed.
}
\examples{
\donttest{
check_req_pkg("data.table")
check_req_pkg(c("base", "utils", "ggplot2", "data.table"))
}

}
