% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-squared test}
\usage{
chi_squared_test(
  data = NULL,
  iv_name = NULL,
  dv_name = NULL,
  sigfigs = 2,
  correct = TRUE
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{iv_name}{name of the independent variable}

\item{dv_name}{name of the dependent variable (must be a binary variable)}

\item{sigfigs}{number of significant digits to round to (for the
table of proportions). By default \code{sigfigs = 2}}

\item{correct}{logical. Should continuity correction be applied?
(default = TRUE)}
}
\description{
Conduct a chi-squared test and produce a contingency table
}
\examples{
chi_squared_test(data = mtcars, iv_name = "cyl", dv_name = "am")
}
