% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_of_mean.R
\name{ci_of_mean}
\alias{ci_of_mean}
\title{Confidence Interval of the Mean of a Vector}
\usage{
ci_of_mean(x = NULL, confidence_level = 0.95, notify_na_count = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{confidence_level}{What is the desired confidence level
expressed as a decimal? (default = 0.95)}
}
\value{
the output will be a named numeric vector with the
lower and upper limit of the confidence interval.
}
\description{
Returns the confidence interval of the mean of a numeric vector.
}
\examples{
ci_of_mean(x = 1:100, confidence_level = 0.95)
ci_of_mean(mtcars$mpg)
}
