% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_of_variation.R
\name{coefficent_of_variation}
\alias{coefficent_of_variation}
\title{Coefficient of variation}
\usage{
coefficent_of_variation(vector, pop_or_sample = "pop")
}
\arguments{
\item{vector}{a numeric vector}

\item{pop_or_sample}{should coefficient of variation be calculated for
a "population" or a "sample"?}
}
\value{
a numeric value
}
\description{
Calculates the (population or sample) coefficient of variation
of a given numeric vector
}
\examples{
coefficent_of_variation(1:4, pop_or_sample = "sample")
coefficent_of_variation(1:4, pop_or_sample = "pop")
}
