% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_d_for_one_sample.R
\name{cohen_d_for_one_sample}
\alias{cohen_d_for_one_sample}
\title{Calculate Cohen's d to accompany a one-sample t-test}
\usage{
cohen_d_for_one_sample(x = NULL, mu = NULL)
}
\arguments{
\item{x}{a numeric vector containing values whose mean will be calculated}

\item{mu}{the true mean}
}
\description{
To run this function, the following package(s) must be installed:
Package 'psych' v2.1.9 (or possibly a higher version) by
William Revelle (2021),
\url{https://cran.r-project.org/package=psych}
}
\examples{
cohen_d_for_one_sample(x = 1:10, mu = 3)
cohen_d_for_one_sample(x = c(1:10, NA, NA), mu = 3)
}
