% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_d_from_cohen_textbook.R
\name{cohen_d_from_cohen_textbook}
\alias{cohen_d_from_cohen_textbook}
\title{Cohen's d from Jacob Cohen's textbook (1988)}
\usage{
cohen_d_from_cohen_textbook(sample_1_as_vector, sample_2_as_vector)
}
\arguments{
\item{sample_1_as_vector}{a vector of values in the first of two samples}

\item{sample_2_as_vector}{a vector of values in the second of two samples}
}
\description{
Calculates Cohen's d as described in Jacob Cohen's textbook (1988),
Statistical Power Analysis for the Behavioral Sciences, 2nd Edition
}
\examples{
cohen_d_from_cohen_textbook(1:10, 3:12)
}
