% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_sep_string_to_numbers.R,
%   R/comma_separated_string_to_numbers.R
\name{comma_sep_string_to_numbers}
\alias{comma_sep_string_to_numbers}
\title{Convert a comma-separated string of numbers}
\usage{
comma_sep_string_to_numbers(string)

comma_sep_string_to_numbers(string)
}
\arguments{
\item{string}{a character string consisting of numbers
separated by commas}
}
\value{
a character string

a character string
}
\description{
Convert a comma-separated string of numbers

Convert a comma-separated string of numbers
}
\examples{
comma_sep_string_to_numbers("1, 2, 3,4,  5  6")
comma_sep_string_to_numbers("1, 2, 3,4,  5  6")
}
