% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{contingency_table}
\alias{contingency_table}
\title{Contingency Table}
\usage{
contingency_table(
  data = NULL,
  dv_name = NULL,
  iv_names = NULL,
  list_of_levels_in_ivs = NULL,
  sigfigs = 3
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{dv_name}{name of the binary dependent variable}

\item{iv_names}{a character vector containing names of the
independent variables}

\item{list_of_levels_in_ivs}{a list of character vectors, each of
which will be used to subset the data and determine the order of
levels in the respective independent variables}

\item{sigfigs}{number of significant digits to which to round
values in the contingency table (default = 2)}
}
\value{
the output will be a contingency table in a data.table format
}
\description{
Create a contingency table.
}
\examples{
contingency_table(
  data = mtcars, dv_name = "am", iv_names = c("cyl", "vs"))
}
