% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_percent_plot.R
\name{cum_percent_plot}
\alias{cum_percent_plot}
\title{Cumulative percentage plot}
\usage{
cum_percent_plot(vector, output_type = "plot")
}
\arguments{
\item{vector}{a numeric vector}

\item{output_type}{if \code{output_type = "plot"}, return a cumulative
percentage plot; if \code{output_type = "dt"}, return a data.table
with cumulative percentages.}
}
\description{
Plots or tabulates cumulative percentages associated with
elements in a vector
}
\examples{
cum_percent_plot(c(1:100, NA, NA))
cum_percent_plot(mtcars$mpg)
cum_percent_plot(vector= mtcars$mpg, output_type = "dt")
}
