% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simplify.R
\name{data_simplify}
\alias{data_simplify}
\title{Simplify data for analysis}
\usage{
data_simplify(
  data = NULL,
  iv_name = NULL,
  dv_name = NULL,
  mod_name = NULL,
  na.omit = TRUE,
  output_type = "dt"
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{iv_name}{name of the independent variable (IV)}

\item{dv_name}{name of the dependent variable (DV)}

\item{mod_name}{name of the moderator variable (MOD)}

\item{na.omit}{logical. If \code{na.omit = TRUE}, rows containing
an NA value in any of the focal variables (e.g., IV, DV, and MOD)
will be removed in the simplified data. By default, \code{
na.omit = TRUE}}
}
\description{
Simplify data for analyses commonly conducted by
}
\examples{
\donttest{
data_simplify(
data = mtcars, iv_name = "vs", dv_name = "mpg", mod_name = "hp")
}
}
