% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats.R
\name{desc_stats}
\alias{desc_stats}
\title{Descriptive statistics}
\usage{
desc_stats(vector = NULL, notify_na_count = NULL)
}
\arguments{
\item{vector}{a numeric vector}

\item{notify_na_count}{if \code{TRUE}, notify how many observations
were removed due to missing values. By default, NA count will be printed
only if there are any NA values.}
}
\value{
a named numeric vector
}
\description{
Returns descriptive statistics for a numeric vector.
}
\examples{
desc_stats(1:100)
desc_stats(c(1:100, NA))
}
