% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats_by_group.R
\name{desc_stats_by_group}
\alias{desc_stats_by_group}
\title{Descriptive statistics by group}
\usage{
desc_stats_by_group(
  data = NULL,
  var_for_stats = NULL,
  grouping_vars = NULL,
  sigfigs = NULL,
  cols_to_round = NULL
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{var_for_stats}{name of the variable for which descriptive
statistics will be calculated}

\item{grouping_vars}{name(s) of grouping variables}

\item{sigfigs}{number of significant digits to round to}

\item{cols_to_round}{names of columns whose values will be rounded}
}
\value{
a data.frame
}
\description{
Returns descriptive statistics by group
}
\examples{
desc_stats_by_group(data = mtcars, var_for_stats = "mpg",
grouping_vars = c("vs", "am"))
}
