% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_formula_convert.R
\name{excel_formula_convert}
\alias{excel_formula_convert}
\title{Excel formula, convert (to)}
\usage{
excel_formula_convert(vector = NULL)
}
\arguments{
\item{vector}{a character vector}
}
\value{
the output will be a character vector formatted as an Excel
formula. For example, if an element in the input vector was \code{".500"},
this element will be converted to \code{=".500"}, which will show up as
".500" in Excel, rather than as "0.5"
}
\description{
Alias for the 'convert_to_excel_formula' function.
Convert elements of a character vector to Excel formulas to preserve
the character (string) format when opened in an Excel file.
}
\examples{
\dontrun{
# compare the two csv files below
# example 1
dt <- data.table::data.table(a = ".500")
data.table::fwrite(dt, "example1.csv") # the csv will show "0.5"
# example 2
dt <- data.table::data.table(a = excel_formula_convert(".500"))
data.table::fwrite(dt, "example2.csv") # the csv will show ".500"
}
}
