% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit_from_parent_function.R
\name{exit_from_parent_function}
\alias{exit_from_parent_function}
\title{Exit from a Parent Function}
\usage{
exit_from_parent_function(
  n = 1,
  silent = FALSE,
  message = "Exiting from a parent function"
)
}
\arguments{
\item{n}{the number of generations to go back (default = 1)}

\item{silent}{logical. If \code{silent = TRUE}, a message will be printed.}

\item{message}{message to print}
}
\description{
Exit from a Parent Function
}
\examples{
fn1 <- function() {
print(1)
print(2)
}
fn1()
fn2 <- function() {
print(1)
exit_from_parent_function()
print(2)
}
fn2()
}
