% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_z_transform.R
\name{fisher_z_transform}
\alias{fisher_z_transform}
\title{Fisher's Z transformation}
\usage{
fisher_z_transform(r = NULL)
}
\arguments{
\item{r}{a (vector of) correlation coefficient(s)}
}
\value{
the output will be a vector of Z values which were transformed
from the given r values.
}
\description{
Perform Fisher's r-to-Z transformation for given correlation coefficient(s).
}
\examples{
fisher_z_transform(0.99)
fisher_z_transform(r = seq(0.1, 0.5, 0.1))
}
