% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holm_adjusted_p_values.R
\name{holm_adjusted_p}
\alias{holm_adjusted_p}
\title{Holm-adjusted p-values}
\usage{
holm_adjusted_p(p = NULL)
}
\arguments{
\item{p}{a numeric vector of p-values}
}
\description{
Adjust a vector of p-values using the method proposed by Holm
}
\details{
See the following reference:
Holm 1979 \url{https://www.jstor.org/stable/4615733}
Manual for the 'p.adjust' function in the 'stats' package
\url{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/p.adjust.html}
}
\examples{
holm_adjusted_p(c(.05, .01))
holm_adjusted_p(c(.05, .05, .05))
}
