% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_across_datasets.R
\name{id_across_datasets}
\alias{id_across_datasets}
\title{ID across datasets}
\usage{
id_across_datasets(
  dt_list = NULL,
  id_col_name = "id",
  id_col_position = "first"
)
}
\arguments{
\item{dt_list}{a list of data.table objects}

\item{id_col_name}{name of the column that will contain ID values}

\item{id_col_position}{position of the newly created ID column.
If \code{id_col_position = "first"}, the new ID column will
be placed as the first column in respective data sets.
If \code{id_col_position = "last"}, the new ID column will
be placed as the last column in respective data sets.}
}
\description{
Create an ID column in each of the data sets. The ID values will
span across the data sets.
}
\examples{
# running the examples below requires importing the data.table package.
kim::prep(data.table)
id_across_datasets(
dt_list = list(setDT(copy(mtcars)), setDT(copy(iris))))
id_across_datasets(
dt_list = list(setDT(copy(mtcars)), setDT(copy(iris)), setDT(copy(women))),
id_col_name = "newly_created_id_col",
id_col_position = "last")
}
