% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenu.R
\name{lenu}
\alias{lenu}
\title{lenu: Length of unique values}
\usage{
lenu(x = NULL)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}
}
\value{
a vector, data frame, or array-like 'x' but with duplicate
elements/rows removed.
}
\description{
Extract unique elements and get the length of those elements
}
\examples{
unique(c(10, 3, 7, 10))
lenu(c(10, 3, 7, 10))
unique(c(10, 3, 7, 10, NA))
lenu(c(10, 3, 7, 10, NA))
lenu(c("b", "z", "b", "a", NA, NA, NA))
}
