% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_odds_ratio_to_d.R
\name{log_odds_ratio_to_d}
\alias{log_odds_ratio_to_d}
\title{Convert log odds ratio to Cohen's d}
\usage{
log_odds_ratio_to_d(log_odds_ratio = NULL, unname = TRUE)
}
\arguments{
\item{log_odds_ratio}{log odds ratio (the input can be a vector of values),
which will be converted to Cohen's d}

\item{unname}{logical. Should the names from the input be removed?
(default = TRUE)}
}
\description{
Convert log odds ratio to Cohen'd (standardized mean difference),
as illustrated in Borenstein et al. (2009, p. 47, ISBN: 978-0-470-05724-7)
}
\examples{
\dontrun{
log_odds_ratio_to_d(log(1))
log_odds_ratio_to_d(log(2))
}
}
