% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu.R
\name{lu}
\alias{lu}
\title{lu: Length of unique values}
\usage{
lu(x = NULL)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}
}
\value{
a vector, data frame, or array-like 'x' but with duplicate
elements/rows removed.
}
\description{
Extract unique elements and get the length of those elements
}
\examples{
unique(c(10, 3, 7, 10))
lu(c(10, 3, 7, 10))
unique(c(10, 3, 7, 10, NA))
lu(c(10, 3, 7, 10, NA))
lu(c("b", "z", "b", "a", NA, NA, NA))
}
