% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_prep_dt.R
\name{matrix_prep_dt}
\alias{matrix_prep_dt}
\title{Prepare a two-column data.table that will be used to fill values in a matrix}
\usage{
matrix_prep_dt(row_var_names = NULL, col_var_names = NULL)
}
\arguments{
\item{row_var_names}{a vector of variable names, each of which will be
header of a row in the eventual matrix}

\item{col_var_names}{a vector of variable names, each of which will be
header of a column in the eventual matrix}
}
\description{
Prepare a two-column data.table that will be used to fill values in a matrix
}
\examples{
matrix_prep_dt(
  row_var_names = c("mpg", "cyl"),
  col_var_names = c("hp", "gear")
)
}
