% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_center.R
\name{mean_center}
\alias{mean_center}
\title{Mean center}
\usage{
mean_center(x)
}
\arguments{
\item{x}{a numeric vector; though not thoroughly tested, the function
can accept a matrix as an input.}
}
\description{
Mean-center a variable, i.e., subtract the mean of a numeric vector
from each value in the numeric vector
}
\examples{
mean_center(1:5)
mean_center(1:6)
# if the input is a matrix
matrix(1:9, nrow = 3)
mean_center(matrix(1:9, nrow = 3))
}
