% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes_of_objects.R
\name{modes_of_objects}
\alias{modes_of_objects}
\title{Find modes of objects}
\usage{
modes_of_objects(...)
}
\arguments{
\item{...}{R objects.}
}
\value{
the output will be a data.table listing objects and their mods.
}
\description{
Find modes of objects
}
\examples{
modes_of_objects(
TRUE, FALSE, 1L, 1:3, 1.1, c(1.2, 1.3), "abc", 1 + 2i, intToBits(1L))
}
