% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentrality_parameter.R
\name{noncentrality_parameter}
\alias{noncentrality_parameter}
\title{Find noncentrality parameter}
\usage{
noncentrality_parameter(t_stat, df, initial_value = 0, ci = 0.95)
}
\arguments{
\item{t_stat}{the t-statistic associated with the noncentrality parameters}

\item{df}{degrees of freedom associated with the noncentrality parameters}

\item{initial_value}{initial value of the noncentrality parameter for
optimization (default = 0). Adjust this value if results look strange.}

\item{ci}{width of the confidence interval associated with the
noncentrality parameters (default = 0.95)}
}
\description{
Find noncentrality parameter
}
\examples{
\donttest{
noncentrality_parameter(4.29, 9)
}
}
