% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_rows_specifically_in_dt.R
\name{order_rows_specifically_in_dt}
\alias{order_rows_specifically_in_dt}
\title{Order rows specifically in a data table}
\usage{
order_rows_specifically_in_dt(
  dt = NULL,
  col_to_order_by = NULL,
  specific_order = NULL
)
}
\arguments{
\item{dt}{a data.table object}

\item{col_to_order_by}{a character value indicating
the name of the column by which to order the data.table}

\item{specific_order}{a vector indicating a specific order of
the values in the column by which to order the data.table.}
}
\value{
a data.table object whose rows will be ordered as specified
}
\description{
Order rows in a data.table in a specific order
}
\examples{
order_rows_specifically_in_dt(mtcars, "carb", c(3, 2, 1, 4, 8, 6))
}
