% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p0.R
\name{p0}
\alias{p0}
\title{Paste0}
\usage{
p0(..., collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.
This is the same argument that would be used in the \code{paste0} function.}

\item{collapse}{an optional character string to separate the results.
Not NA_character_.
This is the same argument that would be used in the \code{paste0} function.}

\item{recycle0}{logical indicating if zero-length character
arguments should lead to the zero-length character(0)
after the sep-phase (which turns into "" in the
collapse-phase, i.e., when collapse is not NULL).
This is the same argument that would be used in the \code{paste0} function.}
}
\description{
A shorthand for the function \code{paste0}
Concatenate vectors after converting to character.
}
\examples{
paste0("a", "b")
p0("a", "b")
}
