% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_list_default.R
\name{package_list_default}
\alias{package_list_default}
\title{Packages - List the default packages}
\usage{
package_list_default(package_type = c("base", "recommended"))
}
\arguments{
\item{package_type}{a vector of package types. By default,
\code{package_type = c("base", "recommended")}}
}
\description{
List the default packages in R
}
\examples{
package_list_default()
package_list_default(package_type = "base")
}
