% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_rank.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Percentile rank}
\usage{
percentile_rank(vector)
}
\arguments{
\item{vector}{a numeric vector}
}
\description{
Calculate percentile rank of each value in a vector
}
\examples{
percentile_rank(1:5)
percentile_rank(1:10)
percentile_rank(1:100)
}
