% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{pm}
\alias{pm}
\title{Paste for message}
\usage{
pm(..., collapse = NULL)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.
Input(s) to this argument will be passed onto the paste0 function.}

\item{collapse}{an optional character string to separate the results.
Not \code{NA_character_}. Input(s) to this argument will be passed onto
the paste0 function.}
}
\value{
there will be no output from this function. Rather, a message
will be generated from the arguments.
}
\description{
Combines the base functions paste0 and message
}
\examples{
pm("hello", 123)
pm(c("hello", 123), collapse = ", ")
}
