% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_variance.R
\name{population_variance}
\alias{population_variance}
\title{Population variance of a vector}
\usage{
population_variance(vector, na.rm = TRUE)
}
\arguments{
\item{vector}{a numeric vector}

\item{na.rm}{if \code{TRUE}, NA values will be removed before calculation}
}
\description{
Calculates the population variance, rather than the sample variance,
of a vector
}
\examples{
population_variance(1:4)
var(1:4)
}
