% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Prepare package(s) for use}
\usage{
prep(...)
}
\arguments{
\item{...}{names of packages to load, separated by commas,
e.g., \code{"ggplot2", data.table}. The arguments can be any number
of packages, and they may or may not be wrapped in quotes.}
}
\description{
Loads package(s). If package(s) are not installed,
install them before loading.
}
\examples{
\donttest{
prep(data.table)
prep("base", utils, ggplot2, "data.table")
}

}
