% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Read a csv file}
\usage{
read_csv(name = NULL, head = FALSE)
}
\arguments{
\item{name}{a character string of the csv file name.
For example, if the csv file to read is "myfile.csv",
enter \code{name = "myfile"}}

\item{head}{logical. if \code{head = TRUE}, prints the first five
rows of the data set.}
}
\value{
the output will be a data.table object, that is,
an output from the data.table function, \code{fread}
}
\description{
Read a csv file
}
\examples{
\donttest{
mydata <- read_csv("myfile")
}
}
