% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_match.R
\name{regex_match}
\alias{regex_match}
\title{Regular expression matches}
\usage{
regex_match(regex = NULL, vector = NULL, mute_report = FALSE, perl = FALSE)
}
\arguments{
\item{regex}{a regular expression
provided, a default theme will be used.}

\item{vector}{a character vector in which to search for regular expression
matches}

\item{mute_report}{logical. Should the report on regular expression
matches be printed?}

\item{perl}{logical. Should Perl-compatible regexps be used?}
}
\description{
Returns elements of a character vector that match the given
regular expression
}
\examples{
regex_match("p$", names(mtcars))
\donttest{
colnames_ending_with_p <- regex_match("p$", names(mtcars))
}
}
