% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_pos_of_value_in_vector.R
\name{rel_pos_of_value_in_vector}
\alias{rel_pos_of_value_in_vector}
\title{Find relative position of a value in a vector}
\usage{
rel_pos_of_value_in_vector(value = NULL, vector = NULL)
}
\arguments{
\item{value}{a value whose relative position is to be searched in a vector}

\item{vector}{a numeric vector}
}
\value{
a number indicating the relative position of the value in the
vector
}
\description{
Find relative position of a value in a vector that may or may not contain
the value
}
\examples{
rel_pos_of_value_in_vector(value = 3, vector = c(2, 4))
rel_pos_of_value_in_vector(value = 3, vector = c(2, 6))
rel_pos_of_value_in_vector(value = 3, vector = 1:3)
}
