% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_value_of_pos_in_vector.R
\name{rel_value_of_pos_in_vector}
\alias{rel_value_of_pos_in_vector}
\title{Find relative value of a position in a vector}
\usage{
rel_value_of_pos_in_vector(vector = NULL, position = NULL)
}
\arguments{
\item{vector}{a numeric vector}

\item{position}{position of a vector}
}
\value{
a number indicating the relative value of the position in the
vector
}
\description{
Find relative value of a position in a vector
}
\examples{
rel_value_of_pos_in_vector(vector = c(0, 100), position = 1.5)
rel_value_of_pos_in_vector(vector = 2:4, position = 2)
rel_value_of_pos_in_vector(vector = c(2, 4, 6), position = 2.5)
}
