% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_user_installed_pkgs.R
\name{remove_user_installed_pkgs}
\alias{remove_user_installed_pkgs}
\title{Remove all user installed packages}
\usage{
remove_user_installed_pkgs(
  exceptions = NULL,
  type_of_pkg_to_keep = c("base", "recommended"),
  keep_kim = FALSE
)
}
\arguments{
\item{exceptions}{a character vector of names of packages to keep}

\item{type_of_pkg_to_keep}{a character vector indicating types
of packages to keep. The default,
\code{type_of_pkg_to_keep = c("base", "recommended")}, keeps all
base and recommended packages that come with R when R is installed.}

\item{keep_kim}{logical. If \code{keep_kim = FALSE}, Package 'kim'
will be removed along with all other user-installed packages.
If \code{keep_kim = TRUE}, Package 'kim' will not be removed.
By default, \code{keep_kim = FALSE}}
}
\description{
Remove all user installed packages
}
\examples{
\dontrun{
remove_user_installed_pkgs()
}
}
