% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_regression.R
\name{robust_regression}
\alias{robust_regression}
\title{Robust regression (bootstrapped regression)}
\usage{
robust_regression(
  data = NULL,
  formula = NULL,
  sigfigs = NULL,
  round_to_nth_digit_after_decimal = NULL,
  iterations = 1000
)
}
\arguments{
\item{data}{a data.frame or data.table object}

\item{formula}{a formula object for the regression equation}

\item{sigfigs}{number of significant digits to round to}

\item{round_to_nth_digit_after_decimal}{round to nth digit after decimal
(alternative to \code{sigfigs})}

\item{iterations}{number of bootstrap samples. The default is set at 1000,
but larger values like 5000 or 10000 are recommended if the user
can tolerate slower handling time.}
}
\description{
Estimate coefficients in a multiple regression model by bootstrapping
Bootstrapping allo
}
\examples{
robust_regression(data = mtcars, formula = mpg ~ cyl * hp)
}
