% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_of_mean.R
\name{se_of_mean}
\alias{se_of_mean}
\title{Standard error of the mean}
\usage{
se_of_mean(vector, na.rm = TRUE, notify_na_count = NULL)
}
\arguments{
\item{vector}{a numeric vector}

\item{na.rm}{if \code{TRUE}, NA values will be removed before calculation}

\item{notify_na_count}{if \code{TRUE}, notify how many observations
were removed due to missing values.}
}
\description{
Standard error of the mean
}
\examples{
se_of_mean(c(1:10, NA))
}
