% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_of_percentage.R
\name{se_of_percentage}
\alias{se_of_percentage}
\title{Standard Error (SE) of a percentage}
\usage{
se_of_percentage(percent = NULL, n = NULL)
}
\arguments{
\item{percent}{percentage; a value between 0 and 100}

\item{n}{sample size; number of observations used to calculate
the percentage}
}
\description{
Calculate the standard error of a percentage.
See Fowler, Jr. (2014, p. 34, ISBN: 978-1-4833-1240-8)
}
\examples{
se_of_percentage(percent = 40, n = 50)
se_of_percentage(percent = 50, n = 10)
}
