% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_of_proportion.R
\name{se_of_proportion}
\alias{se_of_proportion}
\title{Standard Error (SE) of a proportion}
\usage{
se_of_proportion(p = NULL, n = NULL)
}
\arguments{
\item{p}{a vector of proportions; each of the proportion values
must be between 0 and 1}

\item{n}{a vector of sample sizes; number of observations used
to calculate each of the percentage values}
}
\description{
Calculate the standard error of a proportion.
See Anderson and Finn (1996, p. 364, ISBN: 978-1-4612-8466-6)
}
\examples{
se_of_proportion(p = 0.56, n = 400)
se_of_proportion(p = 0.5, n = 10)
}
