% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_to_active_doc.R
\name{setwd_to_active_doc}
\alias{setwd_to_active_doc}
\title{Set working directory to active document in RStudio}
\usage{
setwd_to_active_doc()
}
\value{
there will be no output from this function. Rather, the
working directory will be set as location of the active document.
}
\description{
Set working directory to location of the active document in RStudio
}
\examples{
\dontrun{
setwd_to_active_doc()
}
}
