% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotlight_2_by_continuous.R
\name{spotlight_2_by_continuous}
\alias{spotlight_2_by_continuous}
\title{Spotlight 2 by Continuous}
\usage{
spotlight_2_by_continuous(
  data = NULL,
  iv_name = NULL,
  dv_name = NULL,
  mod_name = NULL,
  interaction_p_include = TRUE,
  iv_level_order = NULL,
  output_type = "plot",
  colors = c("red", "blue"),
  lines_connecting_means = TRUE,
  line_size = 1,
  dot_size = 3,
  error_bar = "ci",
  error_bar_range = 0.95,
  error_bar_tip_width = 0.13,
  error_bar_thickness = 1,
  position_dodge = 0.13,
  simple_effect_text_part_1 = "Simple Effect\\n",
  simple_effect_color = "black",
  simple_effect_font_size = 5,
  interaction_p_value_font_size = 6,
  x_axis_title = NULL,
  y_axis_title = NULL,
  legend_title = NULL,
  legend_position = "right",
  y_axis_title_vjust = 0.85,
  round_decimals_int_p_value = 3,
  plot_margin = ggplot2::unit(c(60, 7, 7, 7), "pt")
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{iv_name}{name of the binary independent variable}

\item{dv_name}{name of the dependent variable}

\item{mod_name}{name of the continuous moderator variable}

\item{interaction_p_include}{logical. Should the plot include a
p-value for the interaction term?}

\item{iv_level_order}{order of levels in the independent
variable for legend. By default, it will be set as levels of the
independent variable ordered using R's base function \code{sort}.}

\item{output_type}{type of output (default = "plot").}

\item{colors}{set colors for the two levels of the independent variable
(default = \code{colors = c("red", "blue")}).}

\item{lines_connecting_means}{logical. Should lines connecting means
within each group be drawn? (default = TRUE)}

\item{line_size}{thickness of the lines connecting group means,
(default = 1)}

\item{dot_size}{size of the dots indicating group means (default = 3)}

\item{error_bar}{if \code{error_bar = "se"}; error bars will be +/-1
standard error, if \code{error_bar = "ci"} error bars will be a
confidence interval. By default, \code{error_bar = "ci"}.}

\item{error_bar_range}{width of the confidence interval
(default = 0.95 for a 95 percent confidence interval).
This argument will not apply when \code{error_bar = "se"}}

\item{error_bar_tip_width}{graphically, width of the segments
at the end of error bars (default = 0.13)}

\item{error_bar_thickness}{thickness of the error bars (default = 1)}

\item{position_dodge}{by how much should the group means and error bars
be horizontally offset from each other so as not to overlap?
(default = 0.13)}

\item{simple_effect_text_part_1}{The first part of the text for
labeling simple effects.
By default, \code{simple_effect_text_part_1 = "Simple Effect\n"}}

\item{simple_effect_color}{color for the text indicating p-values
of simple effects (default = "black").}

\item{simple_effect_font_size}{font size of the text indicating
p-values of simple effects (default = 5).}

\item{interaction_p_value_font_size}{font size for the interaction
p value (default = 6)}

\item{x_axis_title}{title of the x axis. By default, it will be set
as input for \code{mod_name}. If \code{x_axis_title = FALSE}, it will
be removed.}

\item{y_axis_title}{title of the y axis. By default, it will be set
as input for \code{dv_name}. If \code{y_axis_title = FALSE}, it will
be removed.}

\item{legend_title}{title of the legend. By default, it will be set
as input for \code{iv_name}. If \code{legend_title = FALSE}, it will
be removed.}

\item{legend_position}{position of the legend (default = "right").
If \code{legend_position = "none"}, the legend will be removed.}

\item{y_axis_title_vjust}{position of the y axis title (default = 0.85).
If default is used, \code{y_axis_title_vjust = 0.85}, the y axis title
will be positioned at 85\% of the way up from the bottom of the plot.}

\item{round_decimals_int_p_value}{To how many digits after the
decimal point should the p value for the interaction term be
rounded? (default = 3)}

\item{plot_margin}{margin for the plot
By default \code{plot_margin = ggplot2::unit(c(60, 7, 7, 7), "pt")}}
}
\description{
Conduct a spotlight analysis for a 2 x Continuous design.
See Spiller et al. (2013) \url{doi:10.1509/jmr.12.0420}.
}
\examples{
\donttest{
spotlight_2_by_continuous(
data = mtcars,
iv_name = "am",
dv_name = "mpg",
mod_name = "qsec")
}
}
