% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize}
\usage{
standardize(x = NULL)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the output will be a vector of the standard scores of the input.
}
\description{
Standardize (i.e., normalize, obtain z-scores, or obtain the
standard scores)
}
\examples{
standardize(1:10)
}
