% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su.R
\name{su}
\alias{su}
\title{su: Sorted unique values}
\usage{
su(x = NULL, na.last = TRUE, decreasing = FALSE)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}

\item{na.last}{an argument to be passed onto the 'sort' function
(in base R) for controlling the treatment of NA values.
If \code{na.last = TRUE}, NA values in the data are put last;
if \code{na.last = FALSE}, NA values are put first;
if \code{na.last = NA}, NA values are removed.
By default, \code{na.last = TRUE}}

\item{decreasing}{logical. Should the sort be increasing or decreasing?
An argument to be passed onto the 'sort' function (in base R).
By default, \code{decreasing = FALSE}}
}
\value{
a vector, data frame, or array-like 'x' but with duplicate
elements/rows removed.
}
\description{
Extract unique elements and sort them
}
\examples{
su(c(10, 3, 7, 10, NA))
su(c("b", "z", "b", "a", NA, NA, NA))
}
