% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_median_or_bottom.R
\name{top_median_or_bottom}
\alias{top_median_or_bottom}
\title{Top, median, or bottom}
\usage{
top_median_or_bottom(vector)
}
\arguments{
\item{vector}{a numeric vector}
}
\value{
a character vector indicating whether each element
in a vector belongs to "top", "median", or "bottom"
}
\description{
Indicates whether each value in a vector belongs to top, median, or bottom
}
\examples{
top_median_or_bottom(c(1, 2, 3, NA))
top_median_or_bottom(c(1, 2, 2, NA))
top_median_or_bottom(c(1, 1, 2, NA))
}
