% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_of_log_odds_ratio_to_var_of_d.R
\name{var_of_log_odds_ratio_to_var_of_d}
\alias{var_of_log_odds_ratio_to_var_of_d}
\title{Convert variance of log odds ratio to variance of d}
\usage{
var_of_log_odds_ratio_to_var_of_d(var_of_log_odds_ratio = NULL)
}
\arguments{
\item{var_of_log_odds_ratio}{the variance of a log odds ratio
(the input can be a vector of values)}
}
\description{
Convert the variance of a log odds ratio to the variance of
a Cohen'd (standardized mean difference), as illustrated in
Borenstein et al. (2009, p. 47, ISBN: 978-0-470-05724-7)
}
\examples{
\dontrun{
var_of_log_odds_ratio_to_var_of_d(1)
}
}
