% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_of_percentage.R
\name{var_of_percentage}
\alias{var_of_percentage}
\title{Variance of a percentage}
\usage{
var_of_percentage(percent = NULL, n = NULL)
}
\arguments{
\item{percent}{percentage; a value between 0 and 100}

\item{n}{sample size; number of observations used to calculate
the percentage}
}
\description{
Calculate the variance of a percentage.
See Fowler, Jr. (2014, p. 34, ISBN: 978-1-4833-1240-8)
}
\examples{
var_of_percentage(percent = 40, n = 50)
var_of_percentage(percent = 50, n = 10)
}
