% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_r.R
\name{weighted_mean_r}
\alias{weighted_mean_r}
\title{Weighted mean correlation}
\usage{
weighted_mean_r(r = NULL, n = NULL)
}
\arguments{
\item{r}{a (vector of) correlation coefficient(s)}

\item{n}{a (vector of) sample size(s)}
}
\value{
the output will be a list of  vector of correlation coefficient(s).
}
\description{
Calculate the weighted mean correlation coefficient for a given
correlations and sample sizes.
See Field (2001) \doi{10.1037/1082-989X.6.2.161}.
}
\examples{
weighted_mean_r(r = c(0.2, 0.4), n = c(100, 100))
weighted_mean_r(r = c(0.2, 0.4), n = c(100, 20000))
}
