% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_z.R
\name{weighted_z}
\alias{weighted_z}
\title{Weighted z}
\usage{
weighted_z(z = NULL, n = NULL)
}
\arguments{
\item{z}{a vector of z values}

\item{n}{a vector of sample sizes which will be used to calculate the
weights, which in turn will be used to calculate the weighted z.}
}
\value{
the output will be a weighted z value.
}
\description{
Calculate the weighted z (for calculating weighted mean correlation).
See p. 231 of the book Hedges & Olkin (1985),
Statistical Methods for Meta-Analysis (ISBN: 0123363802).
}
\examples{
weighted_z(1:3, c(100, 200, 300))
weighted_z(z = c(1:3, NA), n = c(100, 200, 300, NA))
}
