% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv.R
\name{write_csv}
\alias{write_csv}
\title{Write to a csv file}
\usage{
write_csv(data = NULL, name = NULL, timestamp = NULL)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{name}{a character string of the csv file name without the
".csv" extension. For example, if the csv file to write to is
"myfile.csv", enter \code{name = "myfile"}}

\item{timestamp}{logical. Should the timestamp be appended to the
file name?}
}
\value{
the output will be a .csv file in the working directory,
that is, an output from the data.table function, \code{fwrite}
}
\description{
Write to a csv file
}
\examples{
\donttest{
write_csv(mtcars, "mtcars_from_write_csv")
}
}
