% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{z score}
\usage{
z_score(x = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. If \code{na.rm = TRUE}, NA values in the vector
will be removed before calculating z-scores (default = TRUE).}
}
\value{
the output will be a vector of z-scores.
}
\description{
Calculate z-scores (i.e., standardize or obtain the standard scores)
}
\examples{
z_score(1:10)
}
