% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_to_r_transform.R
\name{z_to_r_transform}
\alias{z_to_r_transform}
\title{Z to r transformation (Inverse of Fisher's Z transformation)}
\usage{
z_to_r_transform(z = NULL)
}
\arguments{
\item{z}{a (vector of) Z values}
}
\value{
the output will be a vector of correlation coefficient(s) that
are the result(s) of the Z-to-r transformation.
}
\description{
Perform the Z-to-r transformation (i.e., the inverse of Fisher's
r-to-Z transformation) for given Z value(s).
}
\examples{
z_to_r_transform(2.646652)
z_to_r_transform(z = -3:3)
}
