% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-squared test}
\usage{
chi_squared_test(
  data = NULL,
  row_var_name = NULL,
  col_var_name = NULL,
  row = NULL,
  col = NULL,
  sigfigs = 2,
  correct = TRUE
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{row_var_name}{name of the variable whose values will fill the
rows of the contingency table}

\item{col_var_name}{name of the variable whose values will fill the
columns of the contingency table}

\item{row}{a vector whose values will fill the rows of the
contingency table}

\item{col}{a vector whose values will fill the columns of the
contingency table}

\item{sigfigs}{number of significant digits to round to (for the
frequency table). By default \code{sigfigs = 2}}

\item{correct}{logical. Should continuity correction be applied?
(default = TRUE)}
}
\description{
Conduct a chi-squared test and produce a contingency table
}
\examples{
chi_squared_test(data = mtcars, row_var_name = "am", col_var_name = "cyl")
chi_squared_test(row = mtcars$cyl, col = mtcars$am)
chi_squared_test(mtcars, "am", "cyl", output_type = "dt")
}
