% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_d.R
\name{cohen_d}
\alias{cohen_d}
\title{Cohen's d}
\usage{
cohen_d(
  sample_1 = NULL,
  sample_2 = NULL,
  data = NULL,
  iv_name = NULL,
  dv_name = NULL,
  ci_range = 0.95,
  output_type = "all"
)
}
\arguments{
\item{sample_1}{a vector of values in the first of two samples}

\item{sample_2}{a vector of values in the second of two samples}

\item{data}{a data object (a data frame or a data.table)}

\item{iv_name}{name of the independent variable}

\item{dv_name}{name of the dependent variable}

\item{ci_range}{range of the confidence interval for Cohen's d
(default = 0.95)}

\item{output_type}{#' If \code{output_type == "all"} or
\code{output_type == "d_se_and_ci"}, the output will
be a vector of Cohen's d, its SE and confidence interval.
If \code{output_type == "d_and_ci"}, the output will
be a vector of Cohen's d and its confidence interval.
If \code{output_type == "d"}, the output will be Cohen's d.
If \code{output_type == "ci"}, the output will be a vector
of the confidence interval around Cohen's d.
If \code{output_type == "se"}, the output will be the standard error
of Cohen's d.
By default, \code{output_type == "all"}.}
}
\description{
Calculates Cohen's d, its standard error, and confidence interval.
See Borenstein et al. (2009, ISBN: 978-0-470-05724-7).
}
\examples{
\donttest{
cohen_d(sample_1 = 1:10, sample_2 = 3:12)
cohen_d(data = mtcars, iv_name = "vs", dv_name = "mpg", ci_range = 0.99)
}
}
