% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_w_interaction.R
\name{logistic_reg_w_interaction}
\alias{logistic_reg_w_interaction}
\title{Logistic regression with an interaction term}
\usage{
logistic_reg_w_interaction(
  data = NULL,
  dv_name = NULL,
  iv_1_name = NULL,
  iv_2_name = NULL
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{dv_name}{name of the dependent variable (must be a binary variable)}

\item{iv_1_name}{name of the first independent variable}

\item{iv_2_name}{name of the second independent variable}
}
\value{
the output will be a data.table object (summary).
}
\description{
Conduct logistic regression for a model with an interaction between
two predictor variables
}
\examples{
logistic_reg_w_interaction(data = mtcars, dv_name = "vs",
iv_1_name = "mpg", iv_2_name = "am")
}
