% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_anova_2_way.R
\name{mixed_anova_2_way}
\alias{mixed_anova_2_way}
\title{Mixed ANOVA 2-Way (Two-Way Mixed ANOVA)}
\usage{
mixed_anova_2_way(
  data = NULL,
  dv_name = NULL,
  iv_name_bw_group = NULL,
  repeated_measures_col_names = NULL,
  iv_name_bw_group_values = NULL
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{iv_name_bw_group}{name of the between-group independent variable}

\item{repeated_measures_col_names}{names of the columns containing
the repeated measures}

\item{iv_name_bw_group_values}{restrict all analyses to
observations having these values for the between-group
independent variable}
}
\description{
Conduct a two-way mixed analysis of variance (ANOVA).
}
\details{
The following package(s) must be installed prior to running this function:
Package 'afex' v3.0.9 (or possibly a higher version) by
Fox et al. (2020),
\url{https://cran.r-project.org/package=car}
}
\examples{
\donttest{
mixed_anova_2_way(
  data = iris, iv_name_bw_group = "Species",
  repeated_measures_col_names = c("Sepal.Length", "Petal.Length"))
}
}
