% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_regression.R
\name{multiple_regression}
\alias{multiple_regression}
\title{Summarize multiple regression results in a data.table}
\usage{
multiple_regression(
  data = NULL,
  formula = NULL,
  sigfigs = NULL,
  round_digits_after_decimal = NULL
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{formula}{a formula object for the regression equation}

\item{sigfigs}{number of significant digits to round to}

\item{round_digits_after_decimal}{round to nth digit after decimal
(alternative to \code{sigfigs})}
}
\description{
Summarize multiple regression results in a data.table
}
\examples{
multiple_regression(data = mtcars, formula = mpg ~ gear * cyl)
}
