% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_means.R
\name{plot_group_means}
\alias{plot_group_means}
\title{Plot group means}
\usage{
plot_group_means(
  data = NULL,
  dv_name = NULL,
  iv_name = NULL,
  ci = NULL,
  se = NULL,
  pi = NULL,
  ci_range = 0.95,
  pi_range = 0.95,
  line_size = 1,
  dot_size = 3,
  error_bar_width = 0.13,
  position_dodge = 0.13,
  legend_position = "right",
  output = "plot"
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{dv_name}{name of the dependent variable}

\item{iv_name}{name(s) of the independent variable(s).
Up to two independent variables can be supplied.}

\item{ci}{if \code{TRUE}, error bars will be a confidence interval}

\item{se}{if \code{TRUE}, error bars will be +/-1 standard error}

\item{pi}{if \code{TRUE}, error bars will be a prediction interval}

\item{ci_range}{width of the confidence interval
(default = 0.95 for 95 percent confidence interval)}

\item{pi_range}{width of the prediction interval
(default = 0.95 for 95 percent prediction interval)}

\item{line_size}{thickness of the lines connecting group means,
(default = 1)}

\item{dot_size}{size of the dots indicating group means (default = 3)}

\item{error_bar_width}{graphically, width of the segments
at the end of error bars (default = 0.13)}

\item{position_dodge}{by how much should the group means and error bars
be horizontally offset from each other so as not to overlap?
(default = 0.13)}

\item{legend_position}{position of the legend:
"none", "top", "right", "bottom", "left", "none" (default = "right")}

\item{output}{output type: either "plot" or "table"}
}
\value{
by default, the output will be a ggplot object.
If \code{output = "table"}, the output will be a data.table object.
}
\description{
Plot group means
}
\examples{
plot_group_means(data = mtcars, dv_name = "mpg", iv_name = c("vs", "am"))
}
