% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_round_p_value.R
\name{pretty_round_p_value}
\alias{pretty_round_p_value}
\title{Pretty round p-value}
\usage{
pretty_round_p_value(
  p_value_vector = NULL,
  round_digits_after_decimal = 3,
  include_p_equals = FALSE
)
}
\arguments{
\item{p_value_vector}{one number or a numeric vector}

\item{round_digits_after_decimal}{how many digits after the decimal
point should the p-value be rounded to?}

\item{include_p_equals}{if \code{TRUE}, output will be a string of
mathematical expression including "p", e.g., "p < .01" (default = FALSE)}
}
\value{
the output will be a character vector with p values, e.g.,
a vector of strings like "< .001" (or "p < .001").
}
\description{
Round p-values to the desired number of decimals and remove
leading 0s before the decimal.
}
\examples{
pretty_round_p_value(0.00001)
pretty_round_p_value(0.00001, round_digits_after_decimal = 4)
pretty_round_p_value(0.00001, round_digits_after_decimal = 5)
# WARNING: the line of code below adding precision that may be unwarranted
pretty_round_p_value(0.00001, round_digits_after_decimal = 6)
pretty_round_p_value(
  p_value_vector = 0.049,
  round_digits_after_decimal = 2, include_p_equals = FALSE)
pretty_round_p_value(c(0.0015, 0.0014, 0.0009), include_p_equals = TRUE)
}
