% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_of_mean.R
\name{se_of_mean}
\alias{se_of_mean}
\title{Standard error of the mean}
\usage{
se_of_mean(vector, na.rm = TRUE, notify_na_count = NULL)
}
\arguments{
\item{vector}{a numeric vector}

\item{na.rm}{Deprecated. By default, NA values will be removed
before calculation}

\item{notify_na_count}{if \code{TRUE}, notify how many observations
were removed due to missing values. By default, NA count will be printed
only if there are any NA values.}
}
\value{
the output will be a numeric vector of length one,
which will be the standard error of the mean for the given numeric vector.
}
\description{
Standard error of the mean
}
\examples{
se_of_mean(c(1:10, NA))
}
