% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_r_env.R
\name{setup_r_env}
\alias{setup_r_env}
\title{Set up R environment}
\usage{
setup_r_env(
  clear_console = TRUE,
  clear_global_env = TRUE,
  setwd_to_active_doc = TRUE,
  prep_kim = TRUE
)
}
\arguments{
\item{clear_console}{if \code{TRUE}, clear the console (default = TRUE)}

\item{clear_global_env}{if \code{TRUE}, remove all objects in the
global environment (default = TRUE)}

\item{setwd_to_active_doc}{if \code{TRUE}, set the working
directory to the active document in RStudio (default = TRUE)}

\item{prep_kim}{if \code{TRUE}, unload and load the kim package
(default = TRUE)}
}
\description{
Set up R environment by (1) clearing the console; (2) removing all
objects in the global environment; (3) setting the working directory
to the active document (in RStudio only); (4) unloading and
loading the kim package.
}
\examples{
\dontrun{
setup_r_env()
}
}
