% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_kim.R
\name{start_kim}
\alias{start_kim}
\title{Start kim}
\usage{
start_kim(
  update = TRUE,
  setup_r_env = TRUE,
  default_packages = c("data.table", "ggplot2"),
  silent_load_data_table = TRUE
)
}
\arguments{
\item{update}{logical. If \code{update = TRUE}, updates the current
package "kim" by installing the most recent version (probably from Github)
By default, \code{update = TRUE}}

\item{setup_r_env}{logical. If \code{update = TRUE}, runs the function
setup_r_env in the package "kim". Type "?kim::setup_r_env" to learn more.
By default, \code{setup_r_env = TRUE}}

\item{default_packages}{a vector of names of packages to load and attach.
By default, \code{default_packages = c("data.table", "ggplot2")}}

\item{silent_load_data_table}{logical. If
\code{silent_load_data_table = TRUE}, the data.table package will be
loaded and attached silently. If \code{silent_load_data_table = FALSE},
the data.table package will be loaded with its default message.}
}
\description{
Start kim (update kim; attach default packages; set working directory, etc.)
}
\examples{
\dontrun{
start_kim()
start_kim(default_packages = c("dplyr", "ggplot2"))
start_kim(update = TRUE, setup_r_env = FALSE)
}
}
