% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su.R
\name{su}
\alias{su}
\title{su: Sorted unique values}
\usage{
su(x = NULL)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}
}
\value{
a vector, data frame, or array-like 'x' but with duplicate
elements/rows removed.
}
\description{
Extract unique elements and sort them
}
\examples{
su(c(10, 3, 7, 10))
su(c("b", "z", "b", "a"))
}
