% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_pairwise.R
\name{t_test_pairwise}
\alias{t_test_pairwise}
\title{t-tests, pairwise}
\usage{
t_test_pairwise(
  data = NULL,
  iv_name = NULL,
  dv_name = NULL,
  sigfigs = 3,
  cohen_d = TRUE,
  cohen_d_w_ci = TRUE,
  adjust_p = "holm",
  bonferroni = NULL,
  mann_whitney = TRUE,
  t_test_stats = FALSE,
  t_test_df_decimals = 1,
  sd = FALSE,
  round_p = 3,
  anova = TRUE,
  round_f = 2
)
}
\arguments{
\item{data}{a data object (a data frame or a data.table)}

\item{iv_name}{name of the independent variable}

\item{dv_name}{name of the dependent variable}

\item{sigfigs}{number of significant digits to round to}

\item{cohen_d}{if \code{cohen_d = TRUE}, Cohen's d statistics will be
included in the output data.table.}

\item{cohen_d_w_ci}{if \code{cohen_d_w_ci = TRUE},
Cohen's d with 95\% CI will be included in the output data.table.}

\item{adjust_p}{the name of the method to use to adjust p-values.
If \code{adjust_p = "holm"}, the Holm method will be used;
if \code{adjust_p = "bonferroni"}, the Bonferroni method will be used.
By default, \code{adjust_p = "holm"}}

\item{bonferroni}{The use of this argument is deprecated.
Use the 'adjust_p' argument instead.
If \code{bonferroni = TRUE}, Bonferroni tests will be
conducted for t-tests or Mann-Whitney tests.}

\item{mann_whitney}{if \code{TRUE}, Mann-Whitney test results will be
included in the output data.table. If \code{FALSE}, Mann-Whitney
tests will not be performed.}

\item{t_test_stats}{if \code{t_test_stats = TRUE}, t-test statistic
and degrees of freedom will be included in the output data.table.}

\item{t_test_df_decimals}{number of decimals for the degrees of freedom
in t-tests (default = 1)}

\item{sd}{if \code{sd = TRUE}, standard deviations will be
included in the output data.table.}

\item{round_p}{number of decimal places to which to round
p-values (default = 3)}

\item{anova}{Should a one-way ANOVA be conducted and reported?
(default = TRUE)}

\item{round_f}{number of decimal places to which to round
the f statistic (default = 2)}
}
\value{
the output will be a data.table showing results of all
pairwise comparisons between levels of the independent variable.
}
\description{
Conducts a t-test for every possible pairwise comparison
with Holm or Bonferroni correction
}
\examples{
\dontrun{
t_test_pairwise(data = iris, iv_name = "Species", dv_name = "Sepal.Length")
t_test_pairwise(data = iris, iv_name = "Species",
dv_name = "Sepal.Length", t_test_stats = TRUE, sd = TRUE)
t_test_pairwise(data = iris, iv_name = "Species", dv_name = "Sepal.Length",
mann_whitney = FALSE)
}
}
