% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_kim.R
\name{update_kim}
\alias{update_kim}
\title{Update 'kim' package}
\usage{
update_kim(source = "github")
}
\arguments{
\item{source}{location of the most recent version of the package
(default = "github")}
}
\description{
Updates the current package 'kim' by installing the most recent version
}
\examples{
\donttest{update_kim()}

}
