% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_kim.R
\name{update_kim}
\alias{update_kim}
\title{Update the package 'kim'}
\usage{
update_kim(confirm = TRUE)
}
\arguments{
\item{confirm}{logical. If \code{confirm = TRUE}, the user will
need to confirm the update. If \code{confirm = FALSE}, the confirmation
step will be skipped. By default, \code{confirm = TRUE}.}
}
\value{
there will be no output from this function. Rather, executing
this function will update the current 'kim' package by installing
the most recent version of the package from GitHub.
}
\description{
Updates the current package 'kim' by installing the
most recent version of the package from GitHub
}
\examples{
\dontrun{
if (interactive()) {update_kim()}
}

}
